requirements.txt
clipspy
nltk
nltkdata
